Crate oak_simple_io

Simple I/O driver for communication between the guest and the host via shared memory.

Structs

BufferDescriptor

I/O port descriptor for a buffer.

SimpleIo

The simple I/O channel driver implementation.

Constants

DEFAULT_INPUT_BUFFER

Default I/O ports for the input buffer.

DEFAULT_OUTPUT_BUFFER

Default I/O ports for the output buffer.

INPUT_BUFFER_LENGTH

The length of the buffer that will be used for input messages.

OUTPUT_BUFFER_LENGTH

The length of the buffer that will be used for output messages.

Traits

Translator

Memory address translation function.